+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-### Unreleased
+### 1.1.0
+### 2023-03-21
 
 - Allow using static facade methods
   - You can use `::getFactory()` from a static or non-static context